Code Refactoring and UI modification to make sure they are the same theme as the wholee application
