+++ talent_link/lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:talent_link/widgets/startup_page.dart';
-import 'package:talent_link/widgets/signup_page.dart';
+import 'package:talent_link/widgets/sign_up_widgets/account_created_screen.dart';
+import 'package:talent_link/widgets/applicatin_startup/startup_page.dart';
+import 'package:talent_link/widgets/sign_up_widgets/signup_page.dart';
 
 void main() {
   runApp(const MyApp());
